Switch the MirrorList server to use threading rather than forking
This improves the performance significantly for several reasons:
1. Threading uses a lot less resources than forking in general.
2. Forks have Copy-On-Write for variables. This means that at the
   moment of forking, they will just point to the parent process
   for all of the memory they stored until now, so they don't have
   to actually copy all of the bits of memory over during fork().
   The problem here though is at the moment of metalink update:
   at this moment, the variables that the master proces uses for
   the in-memory structure suddenly all change. This means that
   all of the forks that are still active at that time will have
   to copy all the previous values to make sure they do not get
   impacted by the update. (Linux-specific, I am not sure about
   other Operating Systems on this).

One possible reason I could think threads have been used rather
than forks in the past is that with forking the childs do not
get impacted at the moment of metalink update. However, since
the Operating System has cross-thread locks, this is of no
issue with regards to locking (the OS will take care of it).
The duration of these locks isn't a problem either since the
only change that happens under the lock is the swapping of one
pointer (the old value) to a new value (the just parsed value).

Signed-off-by: Patrick Uiterwijk <puiterwijk@redhat.com>