Do not load fields seldom/never used
The most interesting part of this change is the defer of the host.config
column. If it is not deferred a simple request to

 https://admin.stg.fedoraproject.org/mirrormanager/mirrors

let's the wsgi apache process use more than 900MB in staging.
Unfortunately the memory is also never released. Marking the column
config as deferred the wsgi process does not use more than 70MB. This
change reduces the required memory by more than a factor of 10.

The average time to load this page is also reduced. With deferred it
takes about 2.5 seconds to load the URL above. Without the change it
takes between 12 and 60 seconds to load the page. It is difficult to
compare as the system starts to use swap due to much higher memory
requirements if not deferring the load of that column.

The config column is never used in the whole MirrorManager2 code and
only written during xmlrpc.

Signed-off-by: Adrian Reber <adrian@lisas.de>