get_global_netblocks: use yesterday's date
Fixes #165

The servers providing the IPv4/IPv6 global netblocks do not seem to be
running on UTC and therefore it can happen that when requesting the
netblock list on the first of the month on a UTC system that the
netblock list does not yet exist as the destination server is still in
the previous month. Using yesterday's date should work around/fix this.

Signed-off-by: Adrian Reber <adrian@lisas.de>