mirrorlist: metalink and mirrorlist should work with the same data
This is in preparation of #100 (Add a way to specify you want only https
urls from metalink). Metalink and mirrorlist creation are pretty
similar. Both get a list of host ids and URLs to fill generate the final
result: [(hostid, [url, url, url]], ...]. In the case of the mirrorlist this list
is trimmed to only return one URL per host: [(hostid, url], ...].

The mirrorlist_client.wsgi then expects the URL to be a string and not a
list. The metalink expects a list of URLs.

This changes the protocol trim function to still return a list of URLs,
even if only one URL is returned.

This in preparation to allow the user to specify a certain URL if
desired. The protocol trim function will then be called before creating
the metalink and thus the trim function needs to return a format which
the metalink creation understands (list of URLs). Also the mirrorlist
creation is adapted to expect a list of URLs instead of a string.

Signed-off-by: Adrian Reber <adrian@lisas.de>