mirrolist: fix &redirect=1
The commits

  ed7e45903f381a07395b7960a719934414c7c8f5
  236532af7a46d50930bdc3f9c8386a063d5960cf

which were used to support the '&protocol=' option broke the mirrorlist
option to answer with a HTTP redirect (302). This fixes the redirect
code path to also work with the new data format returned by the
mirrorlist server.

v2: loop over all returned URLs until a URL starting with 'http' is found

Signed-off-by: Adrian Reber <adrian@lisas.de>