Private URLs are now restricted to admins
MirrorManager has the concept of private URLs which should be used
for inter-mirror syncing. Unfortunately they are sometimes confused
with the concept of private mirrors. Private URLs will never show up in
the mirrorlist/metalink and if a private mirror (only for a certain
network or host) marks a URL as private it will not show up.

As the private flag is not reflected anywhere on the web pages it can
only be seen in the database.

With this change a private URL is now also displayed as private in the
host category URL view and only admins can mark URLs as private.

Signed-off-by: Adrian Reber <adrian@lisas.de>