mirrorlist: at least 5 mirrors should be returned for country/continent
If less than three mirrors for a country/continent are found the
mirrorlist is extended by the global mirrorlist. This increases the
value from three to five as there have been problems especially with
countries/continents with a low number of mirrors (Australia).

In addition the minimum number of mirrors threshold can now be specified
on the command-line: -m, --minimum

Important to know is that if a specific country is requested from the
mirrorlist using '&country=AU' described mechanism does not take effect,
only for regular requests (without country) or requests for a specific
IP ('&ip=xxx.xxx.xxx.xxx').

Signed-off-by: Adrian Reber <adrian@lisas.de>