Remove 'Master rsync server Access Control List IPs' section
This adds a configuration option to enable/disable the 'Master rsync
server Access Control List IPs' section in the host template. As Fedora
has never (or at least for a long time not) used this functionality this
disables if by default in the Fedora templates.

It could have been completely removed from the Fedora templates, but as
long as the Fedora templates are the only templates I just conditionally
disabled it.

Signed-off-by: Adrian Reber <adrian@lisas.de>