umdl: fix make_file_details_from_checksums()
The function make_file_details_from_checksums() tries to find files
which contain the checksum of (mostly) ISO images. Since Fedora 22 these
files have switched to the BSD-style checksum format and umdl was not
longer able to parse those files. This fixes umdl to be able to parse
the old style and the new BSD-style checksums.

In addition to the upcoming fullfiletimelist change this function can
now also handle requests in the stat()-less fullfiletimelist mode.

v2:
 - fixes based on jeremycline's review

Signed-off-by: Adrian Reber <adrian@lisas.de>