umdl: add fullfiletimelist-* based master scanning
Disk based scanning all modules of the Fedora master mirror takes hours.
A full scan takes (at least) more than 6 hours and stat()s every
directory in the whole tree. This creates a massive amount of IOPS which
are basically only used to detect that most directories have not
changed.

With the newly created fullfiletimelist-* files for fedora-quick-mirror
(https://pagure.io/quick-fedora-mirror/) this can be changed and a nearly
stat()-less master mirror scan is possible and implemented by this commit.

Important to remember is that the result will always only be as good as
the input files (fullfiletimelist-*).

With this commit umdl detects if a 'fullfiletimelist-*' file exists and
instead of walking the directory tree only the found file is parsed. To
switch back to disk based scanning a new switch is introduced:

  --skip-fullfiletimelist
                        Do not look for a fullfiletimelist-*; actually scan
                        the filesystem

Using the fullfiletimelist-* files umdl is much faster:

 * Fedora Linux
   * disk based scanning: 48m43.167s
   * fullfiletimelist-fedora: 2m5.367s
 * Fedora EPEL
   * disk based scanning: 2m59.480s
   * fullfiletimelist-epel: 0m5.432s
 * Fedora Secondary Arches
   * disk based scanning: 80m13.825s
   * fullfiletimelist-fedora-secondary: 2m51.272s
 * Fedora Archive
   * disk based scanning: 182m49.157s
   * fullfiletimelist-archive: 2m50.970s
 * Fedora Other
   * disk based scanning: 173m24.203s
   * fullfiletimelist-alt: 10m47.376s

So it is much faster but heavily depends on the correctness of the
fullfiletimelist-* files. Currently a single stat() for every changed
directory is necessary to detect if the directory is readable or not
(pre-bitflip scenario mainly). The feature request for
quick-fedora-mirror to include this information already exists:
https://pagure.io/quick-fedora-mirror/issue/40

To read the *-CHECKSUM files for the information about the ISOs disk
access is still required (could also be downloaded via https or rsync)
and also for the checksums of the repomd.xml file for the metalink.

One point which made this whole thing more complicated than necessary is
the 'Fedora Linux' category. All other category topdir point to
/srv/pub/<module> . Only 'Fedora Linux' points to
/srv/pub/<module>/linux -> /srv/pub/fedora/linux .

The fullfiletimelist-fedora, however, starts at /fedora so that the paths
of this single category cannot be joined as easily as for the other
category. This makes the code at some places unnecessarily complicated.

v2:
 - fixes based on jeremycline's review

Signed-off-by: Adrian Reber <adrian@lisas.de>