detect and setup mirrorlist/metalinks for modular Fedora
in the new Modular world we need to setup a new structure for
delivering content. it will look something like bellow

/pub/fedora/linux/modular/
/pub/fedora/linux/modular/development
/pub/fedora/linux/modular/development/bikeshed
/pub/fedora/linux/modular/development/27
/pub/fedora/linux/modular/releases
/pub/fedora/linux/modular/releases/test
/pub/fedora/linux/modular/updates
/pub/fedora/linux/modular/updates/27
/pub/fedora/linux/modular/updates/testing
/pub/fedora/linux/modular/updates/testing/27

under each releasever 27, 28 bikeshed etc will be a directory for
the consumable product Server, Workstation, Atomic, etc. For each
consumable product we will have to setup a new set of prefixes. We
are starting with Server.

Signed-off-by: Dennis Gilmore <dennis@ausil.us>