publiclist: hide disabled arches and products
The original MirrorManager code base had the feature to hide certain
arches and products from the publiclist. The database fields still exist
and are even set to 'false' for certain arches.

There is the ticket https://pagure.io/fedora-infrastructure/issue/5186
which tried to remove unused architectures from the publiclist
(sparc and sparc64). Using the now unused database column 'publiclist'
seems like the right thing to do to hide certain arches (including
source) from the publiclist. RPM Fusion which still uses the original
MirrorManager code base is a good example how it used to look:

 http://mirrors.rpmfusion.org/mm/publiclist/

The architecture 'source' is not displayed, which has in both instances
(Fedora and RPM Fusion) publiclist = False. It still can be seen in
Fedora:

 https://admin.fedoraproject.org/mirrormanager/

I tested my changes in Fedora's staging instance and marked sparc,
sparc64 and ia64 as publiclist=false in the database:

 https://admin.stg.fedoraproject.org/mirrormanager/

Signed-off-by: Adrian Reber <adrian@lisas.de>