run_rsync: fix timeout handling
While writing a test-case for run_rsync() the timeout handling has shown
unexpected behavior. If using a timeout and if the rsync process
finished before the timeout the run_rsync would still wait until the
timeout finished, even if the rsync process has already finished.

This changes the function to either end when rsync ends or to end if the
timeout is reached.

Signed-off-by: Adrian Reber <adrian@lisas.de>