Use "site", "host" and "mirror" consistently
The terms "site" and "host" have a special meaning in MirrorManager. One
'Site' can consist of multiple 'Host's but the terms were not used
consistently. This tries to address that by renaming 'Site name' in the
host to 'Host name' and by renaming 'Mirror' to 'Hosts' when showing the
overview.

Signed-off-by: Adrian Reber <adrian@lisas.de>