crawler: Correctly calculate the remaining time
The calculation for the rsync timeout was completely wrong. Instead of
calculating the remaining time for the current crawler thread the time
the thread was running was calculated. Which was most of the time 1
second for the first crawled category. This lead to rsync being aborted
after 1 second and the crawler switched to http(s).

With this the remaining time of the crawler thread is now calculated
correctly.

Signed-off-by: Adrian Reber <adrian@lisas.de>