Invalidate mirror when toggling private flag
There have been occasions when a mirror was switched from private to
public and offered to users before the crawler verifying that it
actually has all the content it claims to have.

This makes sure that a mirror is invalidated if either the site or the
host private flag is toggled.

Signed-off-by: Adrian Reber <adrian@lisas.de>