Enable MirrorManager2 to be built using Python 3 for Fedora
This changes all the shebangs to an unversioned one that works in
venvs for development, while adjusting the spec file to change all
the shebangs to the correct Python interpreter during the package
build.

Tests are also now run using Python 3 on Fedora, and Python 2 on CentOS 7.

This strategy is similar to what is used for Pagure.

Signed-off-by: Neal Gompa <ngompa13@gmail.com>