Fix tests to work with python3
Some tests were failing with python3 because the tests relied on the
ordering of the data in the database. With this commit the failing tests
are no longer sensitive to the order of the data in the database.

This commit also removes some tests related to the atomic/ testdata
which removed in a previous commit.

Signed-off-by: Adrian Reber <adrian@lisas.de>