Added protobuf definition
This adds a protobuf definition to replace the pickle format used to
transfer data between the MirrorManager backend and the mirrorlist
server. Python's pickle format does not work for MirrorManager if the
backend is using Python2 and the frontend Python3. This also makes it
easier to implement parts in another language as it is not Python
dependent anymore.

To compile the actual Python interface use:
$ protoc --python_out=mirrormanager2/lib mirrormanager.proto

Signed-off-by: Adrian Reber <adrian@lisas.de>