Add protobuf to mm2_refresh_mirrorlist_cache
Although the backend code supported creating protobuf files the actual
script to create the input for the mirrorlist servers did not have an
option to create protobuf output.

To avoid changing existing scripts the mm2_refresh_mirrorlist_cache
script still defaults to create the pkl and if the output is set to the
string 'None' and the protobuf output is enabled ('-p') only the protobuf
file is created.

Signed-off-by: Adrian Reber <adrian@lisas.de>