Disable report_mirror for public mirrors
One of the limitations of the current report_mirror implementation is
that it only transmits a list of directories to MirrorManager. That
works most of the times, but a common error scenario we are seeing is
that the crawler detects that a mirror is out of sync but report_mirror
is still running. This leads to the situation that correctly
deactivated mirrors by the crawler are enabled again. This is one of the
most common problems we are seeing in Fedora with mirrors.

For private mirrors we still need report_mirror, but for public mirror
this commit turns the report_mirror interface off.

Signed-off-by: Adrian Reber <adrian@lisas.de>