Only block report_mirror for private Hosts
The initial commit to disable report_mirror for non-private mirrors did
check if the Host and Site was set to private. It is, however, only
necessary that the Host is private. This way a public Site can have
private and public Hosts and only the private Hosts can use
report_mirror.

Fixes: #286

Signed-off-by: Adrian Reber <adrian@lisas.de>