Silence a few warnings during testing
The latest sqlalchemy is not happy with our database setup:

mirrormanager2/lib/__init__.py:491: SAWarning: relationship
'CategoryDirectory.category' will copy column category.id to column
category_directory.category_id, which conflicts with relationship(s):
'Directory.categories' (copies category.id to
category_directory.category_id), 'Category.directories' (copies
category.id to category_directory.category_id). If this is not the
intention, consider if these relationships should be linked with
back_populates, or if viewonly=True should be applied to one or more if
they are read-only. For the less common case that foreign key
constraints are partially overlapping, the orm.foreign() annotation can
be used to isolate the columns that should be written towards.   The
'overlaps' parameter may be used to remove this warning.

Silencing for now.

Signed-off-by: Adrian Reber <adrian@lisas.de>