Add specific check for duplicate cat URLS
previously, if a user tried to create a category URL that already
existed in a category, it was caught as a SQLALCHEMY integrity error,
and the application did not crash. However we just printed the exception
to the logs, and it looked like an error. This commit adds some error
checking to see if the URL already exitst, and informs the user as such.

Resolves: #317

Signed-off-by: Ryan Lerch <rlerch@redhat.com>